# docs/swagger.yaml
basePath: /
definitions:
...
  app.ListRes:
    properties:
      list:
        description: Data list // 数据清单
      pager:
        allOf:
        - $ref: '#/definitions/app.Pager'
        description: Pagination info // 翻页信息
    type: object
  app.Pager:
    properties:
      page:
        description: Page number // 页码
        type: integer
      pageSize:
        description: Page size // 每页数量
        type: integer
      totalRows:
        description: Total rows // 总行数
        type: integer
    type: object
  app.Res:
    properties:
      code:
        type: integer
      context: {}
      data: {}
      details: {}
      message: {}
      pageIndex:
        description: '1-BASED wire value: 0/absent = non-paginated message, n>0 =
          download page n-1 (internal 0-based). 1-based so page 0 stays distinguishable
          from non-paginated messages under proto3 zero-value elision; only set for
          paginated download messages on pv>=2 connections. Client-to-server PageAck.pageIndex
          stays 0-based. // 线上值为 1-BASED：0/缺省 = 非分页消息，n>0 = 下载页 n-1（内部 0-based）。采用
          1-based 是为了让第 0 页在 proto3 零值不编码规则下仍可与非分页消息区分；仅 pv>=2 连接的分页下载消息会设置。客户端→服务端的
          PageAck.pageIndex 保持 0-based'
      path: {}
      status:
        type: boolean
      vault: {}
    type: object
...
  dto.FileDTO:
    properties:
      contentHash:
        description: Content hash // 内容哈希
        type: string
      createdAt:
        description: Created at time // 创建时间
        type: string
      ctime:
        description: Creation timestamp // 创建时间戳
        type: integer
      entryId:
        description: v3 entry UUID (fs_entry.id) // v3 条目 UUID
        type: string
      id:
        description: File ID // 文件 ID
        type: integer
      lastTime:
        description: Updated timestamp // 更新时间戳
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        type: integer
      path:
        description: File path // 文件路径
        type: string
      pathHash:
        description: Path hash // 路径哈希
        type: string
      rename:
        description: Rename flag // 重命名标记
        type: integer
      size:
        description: File size // 文件大小
        type: integer
      updatedAt:
        description: Updated at time // 更新时间
        type: string
    type: object
...
  dto.FolderCreateRequest:
    properties:
      context:
        description: Context // 同步上下文
        example: ctx123
        type: string
      path:
        description: Folder path // 文件夹路径
        example: NewFolder
        type: string
      pathHash:
        description: Path hash // 路径哈希
        example: fhash456
        type: string
      vault:
        description: Vault name // 保险库名称
        example: MyVault
        type: string
    required:
    - path
    - vault
    type: object
  dto.FolderDTO:
    properties:
      createdAt:
        description: Created at time // 创建时间
        type: string
      ctime:
        description: Creation timestamp // 创建时间戳
        type: integer
      lastTime:
        description: Record update timestamp // 记录更新时间戳
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        type: integer
      path:
        description: Folder path // 文件夹路径
        type: string
      pathHash:
        description: Path hash // 路径哈希值
        type: string
      updatedAt:
        description: Updated at time // 更新时间
        type: string
    type: object
  dto.FolderDeleteRequest:
    properties:
      context:
        description: Context // 同步上下文
        example: ctx123
        type: string
      path:
        description: Folder path // 文件夹路径
        example: OldFolder
        type: string
      pathHash:
        description: Path hash // 路径哈希
        example: fhash789
        type: string
      vault:
        description: Vault name // 保险库名称
        example: MyVault
        type: string
    required:
    - path
    - vault
    type: object
  dto.FolderTreeNode:
    properties:
      children:
        description: Child nodes // 子节点
        items:
          $ref: '#/definitions/dto.FolderTreeNode'
        type: array
      fileCount:
        description: File count // 文件数量
        type: integer
      name:
        description: Node name // 节点名称
        type: string
      noteCount:
        description: Note count // 笔记数量
        type: integer
      path:
        description: Node path // 节点路径
        type: string
    type: object
  dto.FolderTreeResponse:
    properties:
      folders:
        description: Folder tree // 文件夹树
        items:
          $ref: '#/definitions/dto.FolderTreeNode'
        type: array
      rootFileCount:
        description: File count in root // 根目录中的文件数量
        type: integer
      rootNoteCount:
        description: Note count in root // 根目录中的笔记数量
        type: integer
    type: object
...
  dto.NoteDTO:
    properties:
      clientName:
        description: Client name // 客户端名称
        type: string
      clientType:
        description: Client type // 客户端类型
        type: string
      clientVersion:
        description: Client version // 客户端版本
        type: string
      content:
        description: Note content // 笔记内容
        type: string
      contentHash:
        description: Content hash // 内容哈希
        type: string
      createdAt:
        description: Created at time // 创建时间
        type: string
      ctime:
        description: Creation timestamp // 创建时间戳
        type: integer
      entryId:
        description: v3 entry UUID (fs_entry.id) // v3 条目 UUID
        type: string
      id:
        description: Note ID // 笔记 ID
        type: integer
      lastTime:
        description: Record update timestamp // 记录更新时间戳
        type: integer
      mtime:
        description: Modification timestamp // 修改时间戳
        type: integer
      path:
        description: Note path // 笔记路径
        type: string
      pathHash:
        description: Path hash // 路径哈希
        type: string
      size:
        description: Note size // 笔记大小
        type: integer
      updatedAt:
        description: Updated at time // 更新时间
        type: string
      version:
        description: Version number // 版本号
        type: integer
    type: object
...
host: localhost:9000
info:
  contact:
    email: haierkeys@gmail.com
    name: Haierkeys
    url: https://github.com/haierkeys
  description: This is the Fast Note Sync Service HTTP API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Fast Note Sync Service HTTP API
  version: "1.0"
paths:
...
  /api/folder:
    delete:
      consumes:
      - application/json
      description: Soft delete a folder by path or pathHash
      parameters:
      - description: Delete Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.FolderDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/app.Res'
      security:
      - UserAuthToken: []
      summary: Delete folder
      tags:
      - Folder
    get:
      description: Get folder info for current user by path or pathHash
      parameters:
      - description: Folder path // 文件夹路径
        example: Projects/Work
        in: query
        name: path
        type: string
      - description: Path hash // 路径哈希
        example: fhash123
        in: query
        name: pathHash
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.FolderDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Get folder info
      tags:
      - Folder
    post:
      consumes:
      - application/json
      description: Create a new folder or restore a deleted one by path
      parameters:
      - description: Create Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/dto.FolderCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.FolderDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Create folder
      tags:
      - Folder
  /api/folder/files:
    get:
      description: List non-deleted files in a specific folder with pagination and
        sorting
      parameters:
      - description: Folder path // 文件夹路径
        example: Projects
        in: query
        name: path
        type: string
      - description: Path hash // 路径哈希
        example: fhash123
        in: query
        name: pathHash
        type: string
      - description: Sort by field // 排序字段
        example: mtime
        in: query
        name: sortBy
        type: string
      - description: Sort order // 排序顺序
        example: desc
        in: query
        name: sortOrder
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      - description: Page number // 页码
        in: query
        name: page
        type: integer
      - description: Page size // 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.FileDTO'
                        type: array
                    type: object
              type: object
      security:
      - UserAuthToken: []
      summary: List files in folder
      tags:
      - Folder
  /api/folder/notes:
    get:
      description: List non-deleted notes in a specific folder with pagination and
        sorting
      parameters:
      - description: Folder path // 文件夹路径
        example: Projects
        in: query
        name: path
        type: string
      - description: Path hash // 路径哈希
        example: fhash123
        in: query
        name: pathHash
        type: string
      - description: Sort by field // 排序字段
        example: mtime
        in: query
        name: sortBy
        type: string
      - description: Sort order // 排序顺序
        example: desc
        in: query
        name: sortOrder
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      - description: Page number // 页码
        in: query
        name: page
        type: integer
      - description: Page size // 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/app.ListRes'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.NoteDTO'
                        type: array
                    type: object
              type: object
      security:
      - UserAuthToken: []
      summary: List notes in folder
      tags:
      - Folder
  /api/folder/tree:
    get:
      description: Get the complete folder tree structure for a vault
      parameters:
      - description: Tree depth // 树深度
        example: 3
        in: query
        name: depth
        type: integer
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.FolderTreeResponse'
              type: object
      security:
      - UserAuthToken: []
      summary: Get folder tree
      tags:
      - Folder
  /api/folders:
    get:
      description: Get folder list for current user by parent path or pathHash
      parameters:
      - description: Folder path // 文件夹路径
        example: Projects
        in: query
        name: path
        type: string
      - description: Path hash // 路径哈希
        example: fhash123
        in: query
        name: pathHash
        type: string
      - description: Vault name // 保险库名称
        example: MyVault
        in: query
        name: vault
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.FolderDTO'
                  type: array
              type: object
      security:
      - UserAuthToken: []
      summary: Get folder list
      tags:
      - Folder
...
securityDefinitions:
  ShareAuthToken:
    in: header
    name: Share-Token
    type: apiKey
  UserAuthToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"